--- openstack/dcs/v1/maintainwindows/urls.go
@@ -1,16 +1,0 @@
-package maintainwindows
-
-import (
-	"strings"
-
-	"github.com/opentelekomcloud/gophertelekomcloud"
-)
-
-// endpoint/instances/maintain-windows
-const resourcePath = "instances/maintain-windows"
-
-// getURL will build the get url of get function
-func getURL(client *golangsdk.ServiceClient) string {
-	// remove projectid from endpoint
-	return strings.Replace(client.ServiceURL(resourcePath), "/"+client.ProjectID, "", -1)
-}
