--- openstack/dcs/v1/products/requests.go
@@ -1,11 +1,0 @@
-package products
-
-import (
-	"github.com/opentelekomcloud/gophertelekomcloud"
-)
-
-// Get products
-func Get(client *golangsdk.ServiceClient) (r GetResult) {
-	_, r.Err = client.Get(getURL(client), &r.Body, nil)
-	return
-}
