--- openstack/dcs/v1/products/urls.go
@@ -1,16 +1,0 @@
-package products
-
-import (
-	"strings"
-
-	"github.com/opentelekomcloud/gophertelekomcloud"
-)
-
-// endpoint/products
-const resourcePath = "products"
-
-// getURL will build the get url of get function
-func getURL(client *golangsdk.ServiceClient) string {
-	// remove projectid from endpoint
-	return strings.Replace(client.ServiceURL(resourcePath), "/"+client.ProjectID, "", -1)
-}
